pyproject.toml
[project]
name = "cloud-run-python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.125.0",
    "google-cloud-logging>=3.13.0",
    "invoke>=2.2.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = ["jupyter>=1.1.1", "ruff>=0.14.10", "types-invoke>=2.0.0.10"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D103", # Missing docstring in public function
]

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
disallow_untyped_decorators = false
ignore_missing_imports = true
